Moved placeDialog test
